+++ test/range_list_test.dart
@@ -98,13 +98,29 @@ void testReversed() {
 
 void testSubtract() {
   var result = rng(0, 5).subtract(rng(2, 4));
-  expect(result, [[0, 1], [5]], reason: "RangeList.subtract");
+  expect(
+      result,
+      [
+        [0, 1],
+        [5]
+      ],
+      reason: "RangeList.subtract");
   result = rng(0, 5).subtract(rng(0, 5));
   expect(result, [], reason: "RangeList.subtract");
   result = rng(2, 7).subtract(rng(1, 4));
-  expect(result, [[5, 6, 7]], reason: "RangeList.subtract");
+  expect(
+      result,
+      [
+        [5, 6, 7]
+      ],
+      reason: "RangeList.subtract");
   result = rng(2, 7).subtract(rng(4, 9));
-  expect(result, [[2, 3]], reason: "RangeList.subtract");
+  expect(
+      result,
+      [
+        [2, 3]
+      ],
+      reason: "RangeList.subtract");
 }
 
 RangeList rng(int start, int end) {